fix: Refactor DTO structure for socket API

+++ backend/src/main/java/com/ssafy/mimo/socket/global/dto/HubConnectionResponseDto.java
@@ -6,5 +6,5 @@ import lombok.Builder;
 public record HubConnectionResponseDto(
         String type,
         String requestName,
-        Integer hubId
+        Long hubId
 ) {}
